+++ examples/js.imports.js
@@ -1,23 +1,20 @@
-/* eslint-disable import/extensions */
-/* eslint-disable import/no-unresolved */
-/* eslint-disable unused-imports/no-unused-imports */
-/* eslint-disable import/no-extraneous-dependencies */
-/* eslint-disable node/no-missing-import */
-/* eslint-disable node/no-extraneous-import */
-import SvgoChevronRed from '@core/assets/icons/chevron-up.svg';
+/* eslint-disable no-restricted-imports */
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import { a, b } from 'foo.js';
-import { aaa, capitalize, group, sort, v, w, x, y } from 'radash';
+import { aaa, capitalize, group, sort, v, w, x, xx, y } from 'radash';
 import z from 'zoo.js';
 
-import SvgoChevronUp from '~/api/assets/icons/chevron-up.svg';
+import type { lazy } from '~/core/utils/inflect.js';
 
+import SvgoChevronRed from '@@/core/assets/icons/asdfasfasfasfsdaf/asdfasfasdfasf/asdfasdfasfasdfasdf/sdfasdfasfasdf/chevron-up.svg';
 import SvgoChevronGreen from '@/core/assets/icons/chevron-up.svg';
-import SVGO_CHEVRON_DOWN from '~/core/assets/icons/chevron-down.svg';
+import { run } from '#core/utils/inflect.js';
+import SvgoChevronUp from '~/api/assets/icons/chevron-up.svg';
 import { inflect } from '~/core/utils/inflect.js';
-
 import { useAuthStore } from '~/features/auth/stores/use-auth-store';
+import SVGO_CHEVRON_DOWN from '~~/core/assets/icons/chevron-down.svg';
 
 import { useBacklogTasksActionsAPI } from '../api/use-backlog-tasks-actions-api';
 import { useBacklogTasksAPI } from '../api/use-backlog-tasks-api.js';
-
 import * as foo from './foo.js';
+